server/requirements.txt
anyio==3.6.2
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
